+++ lib/shared/src/index.ts
@@ -117,9 +117,6 @@ export {
     uriExtname,
     uriParseNameAndExtension,
     SUPPORTED_URI_SCHEMAS,
-    createRemoteFileURI,
-    isRemoteFileURI,
-    parseRemoteFileURI,
     type FileURI,
 } from './common/uri'
 export { NoopEditor } from './editor'
@@ -155,6 +152,7 @@ export {
     type CodeCompletionsParams,
     type SerializedCodeCompletionsParams,
     type CompletionResponseGenerator,
+    type CompletionResponseWithMetaData,
 } from './inferenceClient/misc'
 export type {
     IndexedKeywordContextFetcher,