+++ lib/shared/src/llm-providers/ollama/completions-client.ts
@@ -1,4 +1,5 @@
-import type { OllamaGenerateErrorResponse, OllamaGenerateParams, OllamaGenerateResponse } from '.'
+import { Ollama } from 'ollama/browser'
+import type { OllamaGenerateParams, OllamaGenerateResponse } from '.'
 import { contextFiltersProvider } from '../../cody-ignore/context-filters-provider'
 import { isDefined } from '../../common'
 import type { OllamaOptions } from '../../configuration'
@@ -10,14 +11,10 @@ import {
 import type { CompletionLogger } from '../../sourcegraph-api/completions/client'
 import type { CompletionResponse } from '../../sourcegraph-api/completions/types'
 import { isAbortError } from '../../sourcegraph-api/errors'
-import { isNodeResponse } from '../../sourcegraph-api/graphql/client'
 import { isError } from '../../utils'
 
-const RESPONSE_SEPARATOR = /\r?\n/
-
 /**
- * The implementation is based on the `createClient` function from
- * `vscode/src/completions/client.ts` with some duplication.
+ * The client for Ollama's generate endpoint.
  */
 export function createOllamaClient(
     ollamaOptions: OllamaOptions,
@@ -28,84 +25,57 @@ export function createOllamaClient(
         params: OllamaGenerateParams,
         abortController: AbortController
     ): CompletionResponseGenerator {
-        const url = new URL('/api/generate', ollamaOptions.url).href
+        const url = new URL(ollamaOptions.url).href
+        const ollama = new Ollama({ host: url })
+        const model = params.model
+
         const log = logger?.startCompletion(params, url)
         const { signal } = abortController
 
         try {
-            const response = await fetch(url, {
-                method: 'POST',
-                body: JSON.stringify({
-                    ...params,
-                    prompt: await params.prompt.toFilteredString(contextFiltersProvider),
-                }),
-                headers: {
-                    'Content-Type': 'application/json',
-                    Connection: 'keep-alive',
-                },
-                signal,
+            const prompt = await params.prompt.toFilteredString(contextFiltersProvider)
+
+            const res = await ollama.generate({
+                model,
+                prompt,
+                options: params.options,
+                stream: true,
             })
 
-            if (!response.ok) {
-                const errorResponse = (await response.json()) as OllamaGenerateErrorResponse
-                throw new Error(`ollama generation error: ${errorResponse?.error || 'unknown error'}`)
+            const completionResponse: CompletionResponse = {
+                completion: '',
+                stopReason: CompletionStopReason.StreamingChunk,
             }
 
-            if (!response.body) {
-                throw new Error('no response body')
-            }
+            for await (const line of res) {
+                if (line.response) {
+                    completionResponse.completion += line.response
+                    yield {
+                        completionResponse,
+                    }
+                }
 
-            const iterableBody = isNodeResponse(response)
-                ? response.body
-                : browserResponseToAsyncIterable(response.body)
-
-            let insertText = ''
-            let stopReason = ''
-            let buffer = ''
-
-            for await (const chunk of iterableBody) {
-                if (signal.aborted) {
-                    stopReason = CompletionStopReason.RequestAborted
+                if (signal?.aborted) {
+                    completionResponse.stopReason = CompletionStopReason.RequestAborted
+                    ollama.abort()
                     break
                 }
 
-                buffer += chunk.toString()
+                if (line.done) {
+                    completionResponse.stopReason = CompletionStopReason.RequestFinished
 
-                for (const chunkString of buffer.split(RESPONSE_SEPARATOR).filter(Boolean)) {
-                    let line: OllamaGenerateResponse
-
-                    try {
-                        line = JSON.parse(chunkString) as OllamaGenerateResponse
-                        buffer = ''
-                    } catch (error) {
-                        // ignore JSON.parse() errors most probably caused by incomplete chunk and wait for the next one.
-                        break
-                    }
-
-                    if (line.response) {
-                        insertText += line.response
-                        yield {
-                            completion: insertText,
-                            stopReason: CompletionStopReason.StreamingChunk,
-                        }
-                    }
-
-                    if (line.done && line.total_duration) {
-                        const timingInfo = formatOllamaTimingInfo(line)
+                    if (line.total_duration) {
                         // TODO(valery): yield debug message with timing info to a tracer
+                        const timingInfo = formatOllamaTimingInfo(line)
                         logDebug?.('ollama', 'generation done', timingInfo.join(' '))
                     }
                 }
             }
 
-            const completionResponse: CompletionResponse = {
-                completion: insertText,
-                stopReason: stopReason || CompletionStopReason.RequestFinished,
-            }
-
+            completionResponse.stopReason = CompletionStopReason.RequestFinished
             log?.onComplete(completionResponse)
 
-            return completionResponse
+            return { completionResponse }
         } catch (error) {
             if (!isAbortError(error) && isError(error)) {
                 log?.onError(error.message, error)
@@ -155,26 +125,4 @@ function formatOllamaTimingInfo(response: OllamaGenerateResponse): string[] {
             : null
 
     return [...formattedMetrics, promptEvalSpeed, responseEvalSpeed].filter(isDefined)
-}
-
-function browserResponseToAsyncIterable(body: ReadableStream<Uint8Array>): {
-    [Symbol.asyncIterator]: () => AsyncGenerator<string, string, unknown>
-} {
-    return {
-        [Symbol.asyncIterator]: async function* () {
-            const reader = body.getReader()
-            const decoder = new TextDecoder('utf-8')
-
-            while (true) {
-                const { value, done } = await reader.read()
-                const decoded = decoder.decode(value, { stream: true })
-
-                if (done) {
-                    return decoded
-                }
-
-                yield decoded
-            }
-        },
-    }
 }