+++ web/CHANGELOG.md
@@ -1,4 +1,13 @@
 
+## 0.2.2
+- Fixes remote repository context as you switch between chats
+- Adds support for context ignore for remote repositories files
+- Fixes link rendering in the mention menu/suggestion panel
+
+## 0.2.1 
+
+- Fixes remote files and remote symbols files link
+
 ## 0.2.0
 
 Initial release includes